`xtask`: Outsource utilities to `mod utils`
